requirements.txt
streamlit==1.41.1
scikit-learn==1.6.0
joblib==1.3.2
scipy==1.10.1
numpy==1.24.3
Sastrawi==1.0.1
catboost==1.2.7
sentence_transformers
nltk==3.8.1
